rename: 전략패턴 도메인 패키지 하위로 이동
