+++ src/main/java/step5/domain/MovingStrategy.java
@@ -1,4 +1,4 @@
-package step5.strategy;
+package step5.domain;
 
 public interface MovingStrategy {
