+++ src/main/java/step5/domain/RandomDrivingStrategy.java
@@ -1,4 +1,4 @@
-package step5.strategy;
+package step5.domain;
 
 import step5.util.NumberGenerator;
